Arreglando pre-renderizado de la autenticacion
